\name{ccrepe-package}
\alias{ccrepe-package}
\docType{package}
\title{
 A package for analysis of sparse compositional data.\cr
 Allows calculation of the similarity measure nc-score and calculation of compositionality-corrected p-values\cr
 for arbitrary similarity scores (including user-defined) applied to compositional data.
}
\description{
 ccrepe was developed for use with microbial relative abundance data data, which is both sparse and compositional in nature.
}
\details{
\tabular{ll}{
Package: \tab ccrepe\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-18\cr
License: \tab MIT\cr
}
}
\author{
Emma Schwager <emma.schwager@gmail.com>, Craig Bielski<craig.bielski@gmail.com>\cr
Maintainer: Emma Schwager <emma.schwager@gmail.com>,\cr
  Craig Bielski<craig.bielski@gmail.com>,\cr
  George Weingart<george.weingart@gmail.com>
}
