\name{ccrepe-package}
\alias{ccrepe-package}
\docType{package}
\title{
 Analysis of (microbial) sparse compositional data.
}
\description{
 This package allows calculation of the similarity measure nc-score and calculation of compositionality-corrected p-values
 for arbitrary similarity scores (including user-defined) applied to compositional data.
 ccrepe was developed for use with microbial relative abundance data, which is both sparse and compositional in nature.
}
\author{
Emma Schwager <emma.schwager@gmail.com>, Craig
Bielski<craig.bielski@gmail.com>,   George Weingart<george.weingart@gmail.com>\cr
Maintainer: George Weingart<george.weingart@gmail.com>
}
\examples{
packageDescription('ccrepe')
}
