\name{ccrepe}
\alias{ccrepe}
\title{
 Calculates compositionality-corrected p-values and q-values for compositional data using an arbitrary distance metric.
}
\description{
    ccrepe calculates compositionality-corrected p-values and q-values for compositional data by generating
	first a null distribution of the distance metric generated by permutation and renormalization of the data,
	and then by generating an alternative distribution of the distance metric by bootstrap resampling of the data.\cr
	The two distributions are compared using a pooled-variance Z-test to give a compositionality-corrected p-value.
 }
\usage{
 ccrepe(
 x = NA,
 y = NA,
 method = cor,
 method.args = list(),
 min.subj = 20,
 iterations = 1000,
 subset.cols.1 = c(0),				 
 subset.cols.2 = c(0), 
 errthresh = 1e-04,
 verbose = FALSE,
 iterations.gap = 100,
 distributions = NA)
 
}
\arguments{
  \item{x}{
   First dataframe containing the relative abundances in cavity1 : Columns are bugs, rows are samples.\cr
   The first column is considered data\cr 
}
  \item{y}{
   Second dataframe (optional) containing the relative abundances in cavity2: Columns are bugs, rows are samples. \cr
   The first column is considered data. 
   This dataframe will be merged with the first dataframe by Row Number, therefore same rows must pertain to same subjects.
   The matrices are merged by row name (Which defaults to row number).  
}
 
  \item{method}{
   Measurement function, such as cor\cr
   Some properties of the function:\cr
   1.Be able to take either two inputs which are vectors or one input which is either a matrix or a data frame\cr
   2.In the case of two inputs, return a single number\cr
   3.In the case of one input, return a matrix in which the (i,j)th entry is the similarity score for column i and column j in the original matrix\cr
   4.Resulting matrix must be symmetric\cr
   5.The inputs must be named x and y\cr
}
  \item{method.args}{
	A list of arguments for the measurement function.
	For example:  In the case of cor, the following would be acceptable:
	method.args = list(method='spearman',use='complete.obs'	)\cr	
	which is the default for 'cor'. 
}
  \item{min.subj}{
  Minimum number of subjects that must be present (default: 20).
}
  \item{iterations}{
  Reboot iterations (default: 1000).
}
  \item{subset.cols.1}{
  Subset of columns from x to work on.  (Default: c(0) - meaning: Use all columns of x. 
  All the columns of x are used for normalization but calculations are performed only with the requested subset (by default, all columns). 
}
 \item{subset.cols.2}{
  Subset of columns from y to work on.  (Default: c(0) - meaning: Use all columns of y.  
  If applicable (y present), all the columns of y are used for normalization, but calculations are performed only with the requested subset (by default, all columns).  
}
  \item{errthresh}{
	errThresh   (default= 0.0001)  Represents the probability of getting all 0's in a given bootstrapped column.
	If a bug/feature/column has too many 0's, its probability of sampling all 0's will be too large; it is therefore not used.
}
  \item{verbose}{
	Indicator whether the User requested verbose Output  (default:FALSE)
}
 \item{iterations.gap}{
	If output is verbose - number of iterations after issue a status message (Default=100 - displayed only if verbose=TRUE).
}
  \item{distributions}{
	Output Distribution file  (default:NA).
}
}
 
\value{
	Returns a dataframe containing the calculations results and the parameters used.\cr
	Parameters shown:\cr
		min.subj, iterations, errthresh, method, subset.cols.1  (and subset.cols.2 -if applicable)\cr
		iterations.gap and method.args.method('spearman','kendall' or 'pearson')\cr
		Their descriptions are above\cr
	The calculation results consist of:\cr
		cor.matrix, p-values.matrix,  q-values.matrix  
}
\author{
  Emma Schwager <emma.schwager@gmail.com> 
}
\examples{
	data1<-matrix(rlnorm(21*23,1,1),nrow=21)
	ccrepe.results <-ccrepe  (x=data1,iterations=200) 
}